+++ crates/mefikit-py/src/lib.rs
@@ -25,10 +25,7 @@ mod mefipy {
     use super::sel;
 
     #[pymodule_export]
-    use super::pyumesh::PyUMesh;
-
-    #[pymodule_export]
-    use super::pyumesh::intersect_2d2d;
+    use super::pyumesh::{PyOverlayOperation, PyUMesh};
 
     #[pymodule_export]
     use super::pyfield::PyField;